refactor: ServiceAsynchronizer 구조체 생성
- Bank의 비동기 관련 작업을 담당하는 구조체 정의
